+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: (Solana) Add new tokens: BSOL, DRIFT, HAWK, JITOSOL, JSOL, JTO, JUP, MIMO, MNGO, MSOL, SOL, USDC.e, WBTC, and WETH
 - added: (EVM) Add dRPC nodes
+- changed: Add evmscan fallback for `eth_getTransactionReceipt`in `getL1RollupFee`
 
 ## 4.19.0 (2024-08-12)